rgw: Fallback to Host header for bucket name.
RGW should fallback to using the Host header as the bucket name even when it is
NOT a suffix match against the DNS names, or a match against the CNAME rule.

This mirrors AWS S3 behavior for these cases (The AWS S3 does not do any DNS
lookups regardless).

Backport: jewel
Fixes: http://tracker.ceph.com/issues/15975
Signed-off-by: Robin H. Johnson <robin.johnson@dreamhost.com>